Updating based on 8f8f69d7

+++ resource-manager/web/2023-12-01/appserviceenvironments/model_functionsalwaysreadyconfig.go
@@ -4,6 +4,6 @@ package appserviceenvironments
 // Licensed under the MIT License. See NOTICE.txt in the project root for license information.
 
 type FunctionsAlwaysReadyConfig struct {
-	InstanceCount *float64 `json:"instanceCount,omitempty"`
-	Name          *string  `json:"name,omitempty"`
+	InstanceCount *int64  `json:"instanceCount,omitempty"`
+	Name          *string `json:"name,omitempty"`
 }
